Initial commit - Upload first version of functioning code
Commit description:
This commit includes the initial version of the code, which consists of the following files:

index.html: The main HTML file for the application.
style.css: The CSS file for styling the application.
./img: A directory containing image assets.
./js/doodlejump.js: JavaScript code for the Doodle Jump game.
./js/flappybird.js: JavaScript code for the Flappy Bird game.
./js/tictactoe.js: JavaScript code for the Tic Tac Toe game.
./js/script.js: JavaScript code responsible for sliding functionality.
Please note that this initial version is not well designed and will be improved in terms of structure, efficiency, and clarity. Code will be properly commented for better understanding and maintainability. Further enhancements will be made in subsequent commits